fix(desktop): extract composer audience hook under file-size ratchet
Move contextual-audience logic out of MessageComposer so the file no
longer grows past the line budget; format mobile SendMessage unit tests.

(cherry picked from commit 445100d2db5bc5b683cd1a16c809a0ed687491cd)
