+++ desktop/src/features/messages/ui/MessageComposer.tsx
@@ -24,14 +24,8 @@ import { useAttachmentEditing } from "@/features/messages/lib/useAttachmentEditi
 import { useMediaUpload } from "@/features/messages/lib/useMediaUpload";
 import { useMentions } from "@/features/messages/lib/useMentions";
 import { diffAddedMentionPubkeys } from "@/features/messages/lib/threading";
-import { useUnaddressedChannelAgentMode } from "@/features/channels/lib/unaddressedChannelAgentMode";
-import {
-  describeComposerAudienceHint,
-  resolveComposerSendAudience,
-} from "@/features/messages/lib/composerSendAudience";
 import { getPersistentAgentAudienceScope } from "@/features/messages/lib/persistentAgentAudience";
 import { useIdentityQuery } from "@/shared/api/hooks";
-import { normalizePubkey } from "@/shared/lib/pubkey";
 import {
   hasMentionClipboardHtml,
   normalizeMentionClipboardHtml,
@@ -57,11 +51,12 @@ import {
 } from "./MentionAutocomplete";
 import { ComposerDockToolbar } from "./ComposerDockToolbar";
 import { NonMemberMentionDialog } from "./NonMemberMentionDialog";
+import { ComposerAudienceHint } from "./ComposerAudienceHint";
+import { useComposerAgentAudience } from "./useComposerAgentAudience";
 import { useMentionSendFlow } from "./useMentionSendFlow";
 import { usePersistentAgentMentionHydration } from "./usePersistentAgentMentionHydration";
 import { useComposerContentState } from "./useComposerContentState";
 import { useDraftPersistLifecycle } from "./useDraftPersistSnapshot";
-
 import type { MessageComposerProps } from "./MessageComposer.types";
 
 function MessageComposerImpl({
@@ -130,7 +125,6 @@ function MessageComposerImpl({
       : null;
   const effectiveDraftKeyRef = React.useRef(effectiveDraftKey);
   effectiveDraftKeyRef.current = effectiveDraftKey;
-  // Snapshot composer state before edit mode so cancel can restore it.
   const preEditSnapshotRef = React.useRef<{
     content: string;
     pendingImeta: ImetaMedia[];
@@ -293,102 +287,27 @@ function MessageComposerImpl({
     mentions,
     richText,
   });
-  const persistentAudience = persistentMentionHydration.audience;
   const persistentMentionHydrationRef = React.useRef(
     persistentMentionHydration,
   );
   persistentMentionHydrationRef.current = persistentMentionHydration;
-  const { mode: unaddressedMode } = useUnaddressedChannelAgentMode();
 
-  const conversationKind = channelType === "dm" ? "direct" : "channel";
-  const channelMemberPubkeyList = React.useMemo(
-    () => [...mentions.memberPubkeys],
-    [mentions.memberPubkeys],
-  );
-  const verifiedChannelAgentPubkeys = React.useMemo(
-    () => channelMemberPubkeyList.filter((pk) => mentions.isAgentPubkey(pk)),
-    [channelMemberPubkeyList, mentions.isAgentPubkey],
-  );
-  const currentAgentPubkey = React.useMemo(() => {
-    if (conversationKind !== "direct") return null;
-    const agents = verifiedChannelAgentPubkeys.filter(
-      (pk) => pk !== normalizePubkey(ownerPubkey ?? ""),
-    );
-    return agents[0] ?? null;
-  }, [conversationKind, ownerPubkey, verifiedChannelAgentPubkeys]);
-
-  const resolveComposerAudience = React.useCallback(
-    ({
-      explicitMentionPubkeys,
-      explicitAgentPubkeys,
-      messagePosition,
-      threadRootEventId,
-    }: {
-      explicitMentionPubkeys: string[];
-      explicitAgentPubkeys: string[];
-      messagePosition: "top-level" | "in-thread";
-      threadRootEventId: string | null;
-    }) =>
-      resolveComposerSendAudience({
-        conversation: conversationKind,
-        messagePosition,
-        unaddressedMode,
-        keepAddressedAgentsActive: persistentAudience.enabled,
-        explicitMentionPubkeys,
-        explicitAgentPubkeys,
-        currentAgentPubkey,
-        channelMemberPubkeys: channelMemberPubkeyList,
-        verifiedChannelAgentPubkeys,
-        persistentThreadAudience: [...persistentAudience.pubkeys],
-        threadRootEventId,
-        recipientLoadError:
-          !mentions.hasResolvedMembers && conversationKind === "channel",
-      }),
-    [
-      channelMemberPubkeyList,
-      conversationKind,
-      currentAgentPubkey,
-      mentions.hasResolvedMembers,
-      persistentAudience.enabled,
-      persistentAudience.pubkeys,
-      unaddressedMode,
-      verifiedChannelAgentPubkeys,
-    ],
-  );
-
-  const composerAudienceHint = React.useMemo(() => {
-    if (editTarget != null || conversationKind === "direct") return null;
-    const text = richText.getPlainTextAndCursor().text;
-    const explicitMentionPubkeys = mentions.extractMentionPubkeys(text);
-    const explicitAgentPubkeys = explicitMentionPubkeys.filter((pk) =>
-      mentions.isAgentPubkey(pk),
-    );
-    // Reuse the same send-path resolver so the hint cannot drift from p-tags.
-    const decision = resolveComposerAudience({
-      explicitMentionPubkeys,
-      explicitAgentPubkeys,
-      messagePosition: audienceThreadRootId ? "in-thread" : "top-level",
-      threadRootEventId: audienceThreadRootId,
-    });
-    return describeComposerAudienceHint({
-      conversation: conversationKind,
-      unaddressedMode,
-      explicitAgentCount: explicitAgentPubkeys.length,
-      implicitAgentCount:
-        explicitAgentPubkeys.length > 0
-          ? 0
-          : decision.agentAudiencePubkeys.length,
-      retainDraft: decision.retainDraft,
-    });
-  }, [
+  const {
+    composerAudienceHint,
+    audienceGeneration,
+    audienceRevision,
+    resolveComposerAudience,
+    onSuccessfulExplicitAgentAudience,
+    resolvePostSendContent,
+  } = useComposerAgentAudience({
     audienceThreadRootId,
-    conversationKind,
+    channelType,
     editTarget,
     mentions,
-    resolveComposerAudience,
+    ownerPubkey,
+    persistentMentionHydration,
     richText,
-    unaddressedMode,
-  ]);
+  });
 
   const mentionSendFlow = useMentionSendFlow({
     channelId,
@@ -406,18 +325,10 @@ function MessageComposerImpl({
     setIsEmojiPickerOpen,
     setPendingImeta: media.setPendingImeta,
     setSpoileredAttachmentUrls,
-    onSuccessfulExplicitAgentAudience:
-      persistentAudience.enabled && audienceContext && ownerPubkey
-        ? ({ channelId: successfulChannelId, ...promotion }) => {
-            const scope = getPersistentAgentAudienceScope({
-              ownerPubkey,
-              channelId: successfulChannelId,
-              threadRootId: audienceThreadRootId,
-            });
-            persistentAudience.promotePubkeys({ ...promotion, scope });
-          }
-        : undefined,
-    resolvePostSendContent: persistentMentionHydration.resolvePostSendContent,
+    onSuccessfulExplicitAgentAudience: audienceContext
+      ? onSuccessfulExplicitAgentAudience
+      : undefined,
+    resolvePostSendContent,
     resolveComposerAudience,
   });
 
@@ -572,7 +483,6 @@ function MessageComposerImpl({
     [richText.editor, mentions.clearMentions, customEmoji],
   );
 
-  // ── @ mention picker (toolbar button) ───────────────────────────────
   const openMentionPicker = React.useCallback(() => {
     if (!richText.editor) return;
     const { text, cursor } = richText.getPlainTextAndCursor();
@@ -603,11 +513,9 @@ function MessageComposerImpl({
     mentions.updateMentionQuery,
   ]);
 
-  // ── Submit message ──────────────────────────────────────────────────
   const submitMessage = React.useCallback(async () => {
     const trimmed = syncComposerContentFromEditor().trim();
 
-    // Edit mode
     if (editTargetRef.current && onEditSaveRef.current) {
       if (isSendingRef.current || isUploadingRef.current) return;
       const currentPendingImeta = media.pendingImetaRef.current;
@@ -673,7 +581,6 @@ function MessageComposerImpl({
       return;
     }
 
-    // Normal send
     const currentPendingImeta = media.pendingImetaRef.current;
     const hasMedia = currentPendingImeta.length > 0;
     if (
@@ -707,8 +614,8 @@ function MessageComposerImpl({
         ),
         spoileredAttachmentUrls,
         trimmed,
-        audienceGeneration: persistentAudience.generation,
-        audienceRevision: audienceScope ? persistentAudience.revision : null,
+        audienceGeneration,
+        audienceRevision: audienceScope ? audienceRevision : null,
       });
     } finally {
       persistentMentionHydration.endSubmit();
@@ -733,21 +640,15 @@ function MessageComposerImpl({
     onCaptureSendContext,
     onPreparingMentionSendChange,
     audienceScope,
+    audienceGeneration,
+    audienceRevision,
     persistentMentionHydration,
-    persistentAudience.generation,
-    persistentAudience.revision,
   ]);
   submitMessageRef.current = submitMessage;
 
-  // ── Auto-submit on draft send ────────────────────────────────────────────
-  // When `autoSubmitDraftKey` is set (the user clicked "Send message" in the
-  // Drafts panel and confirmed), fire `submitMessage` once after mount so the
-  // draft is sent through the real send path (mention resolution, media, etc.).
-  //
-  // Guard: only fire when the effective draft key matches the trigger so a
-  // stale URL param on a different channel never fires a spurious send.
-  //
-  // Fires at most once per mount (empty dep array after the key check) — the
+  // Auto-submit draft once when Drafts panel confirms send.
+  // Guard: effective draft key must match trigger (no cross-channel fire).
+  // Fires at most once per mount — the
   // `onAutoSubmitComplete` callback clears the trigger before `submitMessage`
   // runs, preventing re-fire on re-render or back-navigation.
   const onAutoSubmitCompleteRef = React.useRef(onAutoSubmitComplete);
@@ -786,7 +687,6 @@ function MessageComposerImpl({
 
   // ── Keyboard handling ───────────────────────────────────────────────
   // Tiptap handles formatting shortcuts (⌘B, ⌘I, etc.) natively.
-  // Plain Enter → submit is now handled inside the Tiptap `submitOnEnter`
   // extension (fires before ProseMirror's splitBlock). This wrapper only
   // handles autocomplete arrow/enter keys and Escape for edit mode.
   const handleEditorKeyDown = React.useCallback(
@@ -1054,14 +954,7 @@ function MessageComposerImpl({
               </div>
             ) : null}
 
-            {composerAudienceHint ? (
-              <p
-                className="mb-2 text-2xs text-muted-foreground"
-                data-testid="composer-agent-audience-hint"
-              >
-                {composerAudienceHint}
-              </p>
-            ) : null}
+            <ComposerAudienceHint hint={composerAudienceHint} />
 
             {(media.pendingImeta.length > 0 || media.isUploading) && (
               <div className="mb-2 flex items-center gap-2">
