+++ mobile/test/features/channels/send_message_provider_test.dart
@@ -25,8 +25,7 @@ void main() {
         addLocalMessage: (_, event) => localMessages.add(event),
         completeLocalMessage: (_, eventId) => completedIds.add(eventId),
         removeLocalMessage: (_, eventId) => removedIds.add(eventId),
-        readUnaddressedMode: () =>
-            UnaddressedChannelAgentMode.mentionsOnly,
+        readUnaddressedMode: () => UnaddressedChannelAgentMode.mentionsOnly,
         fetchAgentDirectory: () async => const [],
         readChannel: (_) => null,
       );
